feat: enhance setup script and implement automatic admin bootstrapping
- Rename 'setup-auth.ts' to 'setup.ts' and expand to a full setup wizard
- Add prompts for AUTH_ENABLED and DEMO_MODE in the setup script
- Implement automatic admin account creation on first startup if none exists
- Add setup instructions and links for multiple OAuth providers
- Update documentation and package.json to reflect the new setup workflow
