+++ GEMINI.md
@@ -131,7 +131,7 @@ npm run prepare
     - RBAC with `admin`, `editor`, and `viewer` roles.
     - `AUTH_ENABLED` toggle in `.env` allows for public or private access.
     - JWT-based stateless session management.
-    - `npm run setup-auth` CLI script for initial provisioning.
+    - `npm run setup` CLI script for initial provisioning and environment configuration.
 - **Security & Sandboxing**:
     - **XSS Prevention**: All dynamic content (URLs, names, issue messages) is HTML-escaped in reports and the dashboard.
     - **Rate Limiting**: Global API limits (100 req/min) and strict scan trigger limits (2 req/min) using `@fastify/rate-limit`.
